requirements.txt
pytorch
opencv-python
Cython
numba
progress
matplotlib
easydict
scipy
numpy
